edge/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
paho-mqtt==1.6.1
requests==2.32.3